Updated the logic.js/index.html
Updated the logic.js to contain a variable called submitButton that selects the pageSwitcher id from the index.html. Then I created an event listener function when you click on the submit button that takes you to the next page. Also index.html has an id attribute in the input tag now for the javascript to select.