Updated the index/blog.html, logic/form/blog.js.
Updated the index.html file to include a new id name for pageswitcher called submission. Updated the blog.html file so that the back button tag has an id called back. Then copied the code from logic.js over to form.js since this is where it will load first, and then also copied the code from form.js to blog.js and changed the variable names so that it functions when the blog.html page loads this way both the buttons work.