Update logic.js
Updated the logic.js so that a variable parses (turns strings in to objects) the key 'blogData' then created a function to getBlogsFromLocalStorage. Then had the function get the elements by id which then takes the variable that parses the strings in the array, and turns the strings into textcontent then I displayed them.