Update blog.html/logic.js files
Updated the blog.html to have new h2 p and span id attribute names so they work with the javascript. Updated the logic.js to make the text content grab more elements from the html so that the elements that are grabbed are grabbed from local storage. Also made a template literal string for the title to test if it populates into the blog. Added some more comments to help out with the rest of the work.