Update styles.css/blog/index.html/blog.js
Updated styles.css to have two ids that show the lightmode and darkmode. Updated the blog/index.html so that the button is connected to the ids.  Updated the blog.js so that it has an event listener that listens to the click of the toggle button to change the body element if the mode is light set to dark when clicked else go to light mode need to fix cannot read properties of null reading addeventlistener.