Update index.html/styles.css/blog/form.js
Updated the index.html to include an id in the body of dark. Updated the styles.css to have the correct syntax for the variables, and added a new id attribute called body#dark. Changed the if else statement by moving around the light and dark mode so that the page starts on dark mode then toggles the button for each style in the css every time the toggle btn is clicked. Then copied over the javascript from the blog.js to the form.js and put it at the bottom so that it works on both the index and the blog.html.